move ListViewTest from test_views_old.py to test_views.py
